Avoid duplication between 64k and 1MB linker scripts (#415)
* Extend linker scripts to define all regions

* Put common ld script parts in a common file

* Move ld output format out of common

* Tweak stage2 ld script

* Cleanup alignment directives

* Add missing section directive

* Fix incorrect EFPT offset in ld script

* Fix vasm not providing natural alignment

* Make zip unsigned char
This has no effect, but figured it made sense for predictability.

* Propagate subset of changes to r68k

* Don't assume that .data LMA is right after text

* Update copyright years

* Round updateflash half_size up
Avoid missing the last byte when programming an odd-size image

* Fix a sign in linker script

Flipped a sign to correctly get the end of BASERAM for calculating UMEM length. This has no effect currently, since `ORIGIN(BASERAM)` is 0, but just in case it does this fixes it.